one ray cast on minimap

+++ libft/sources/control_flow/control_flow.h
@@ -6,7 +6,7 @@
 /*   By: inikulin <inikulin@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/17 10:57:29 by inikulin          #+#    #+#             */
-/*   Updated: 2025/04/11 17:48:42 by inikulin         ###   ########.fr       */
+/*   Updated: 2025/04/20 17:46:34 by inikulin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define CONTROL_FLOW_H
 
 int		ft_if_i(int choice, int a, int b);
-int		ft_if_d(int choice, double a, double b);
+double	ft_if_d(int choice, double a, double b);
 char	ft_if_c(int choice, char a, char b);
 char	*ft_if_s(int choice, char *a, char *b);
 #endif
